+++ src/support/helpers.js
@@ -72,6 +72,15 @@ function scopedAlertsMethods(
   };
 }
 
+function addEventListener(listener, eventName, eventEmitter) {
+  const removeCallback = () => {
+    eventEmitter.removeListener(eventName, listener);
+  };
+  eventEmitter.on(eventName, listener);
+  return removeCallback;
+}
+
 module.exports = {
   scopedAlertsMethods,
+  addEventListener,
 };